Move apps Vite behavior into getVitePlugin
